+++ CHANGELOG.md
@@ -6,10 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.3.1] - 2022-08-01
+## [0.4.0] - 2022-08-05
 - Fix transaction value conversion.
 
 - Fix `tx-hash` calculation.
+
+- GasOracle is now optional and defaults to the provider.
+
+- Fix "already known" error.
+
+- Fix "transaction underpriced" error.
 
 ## [0.3.0] - 2022-08-01
 - Add `rustls` feature.
@@ -26,8 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2021-12-28
 - Initial release
 
-[Unreleased]: https://github.com/cartesi-corp/tx-manager/compare/v0.3.1...HEAD
-[0.3.1]: https://github.com/cartesi-corp/tx-manager/compare/v0.3.0...v0.3.1
+[0.4.0]: https://github.com/cartesi-corp/tx-manager/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/cartesi-corp/tx-manager/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/cartesi-corp/tx-manager/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/cartesi-corp/tx-manager/releases/tag/v0.1.0